+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed 'F3' bug: findNext didn't trigger because the context was not (re)set properly on focus-out from the 'find panel'.
 - Changed all context 'when' clauses to be more 'precise & unique' in relation to other extensions, such as leojs.
 - Added the reStructuredText (rest) .rst language support (syntax coloring and snippets)
+- Fixed at-button panel to refresh after adding a button
 
 ## 1.0.2